@@ -27,22 +27,22 @@
       "type": "file",
       "required": true,
       "title": "KeePass Database",
-      "description": "KeePass Database"
+      "description": "KeePass database to access"
     },
     {
       "name": "keepassxcRootPath",
       "type": "appPicker",
       "required": true,
       "title": "KeePassXC App",
-      "description": "KeePassXC App",
+      "description": "KeePassXC app to use",
       "default": "/Applications/KeePassXC.app"
     },
     {
       "name": "lockAfterInactivity",
       "type": "dropdown",
       "required": false,
-      "title": "Lock Database after Inactivity",
-      "description": "Lock Database after Inactivity",
+      "title": "Database Locking after Inactivity",
+      "description": "Time after which the database will be locked",
       "data": [
         {
           "title": "Never",
@@ -86,27 +86,33 @@
         }
       ],
       "default": "-1"
+    },
+    {
+      "name": "userInterfaceFavicon",
+      "type": "checkbox",
+      "required": false,
+      "title": "User Interface",
+      "description": "Whether to display favicons in the user interface",
+      "label": "Favicon",
+      "default": false
     }
   ],
   "dependencies": {
-    "@raycast/api": "^1.93.2",
+    "@raycast/api": "^1.94.3",
     "@raycast/utils": "^1.19.1",
     "@types/crypto-js": "^4.2.2",
     "crypto-js": "^4.2.0",
     "csv-parse": "^5.6.0",
-    "otpauth": "^9.3.6",
-    "run-applescript": "^6.1.0"
+    "otpauth": "^9.4.0",
+    "run-applescript": "^7.0.0"
   },
   "devDependencies": {
-    "@types/node": "~18.18.2",
-    "@types/react": "^18.3.18",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.10.0",
-    "prettier": "^2.8.8",
-    "react-devtools": "^5.2.0",
-    "typescript": "^4.9.5"
+    "@raycast/eslint-config": "^2.0.4",
+    "@types/node": "^22.14.0",
+    "@types/react": "^19.1.0",
+    "eslint": "^9.23.0",
+    "prettier": "^3.5.3",
+    "typescript": "^5.8.2"
   },
   "scripts": {
     "build": "ray build -e dist",
